+++ src/index.ts
@@ -34,6 +34,9 @@ export * from './components/Forms/Inputs/TextArea'
 export * from './components/Forms/Inputs/constants/InputWrapperStyle'
 export * from './components/Forms/FormHelperText'
 
+/* Hooks for Component */
+export { default as useToast } from './components/Toast/useToast'
+
 /* Layout */
 export * from './layout/GNB'
 export * from './layout/GlobalHeader'
@@ -46,6 +49,14 @@ export * from './layout/Side/SideViewContent'
 export * from './layout/redux'
 export * from './layout/LayoutProvider'
 
+/* Hooks for Layout */
+export { default as useLayoutDispatch } from './layout/hooks/useLayoutDispatch'
+export { default as useLayoutState } from './layout/hooks/useLayoutState'
+export { default as useSideWidth } from './layout/hooks/useSideWidth'
+export { default as useSidePanelHandler } from './layout/hooks/useSidePanelHandler'
+export { default as useSideViewHandler } from './layout/hooks/useSideViewHandler'
+export { default as useHeader } from './layout/hooks/useHeader'
+
 /* Types */
 export type { default as BlocksParserContextType } from './types/BlocksParserContextType'
 export type { UIComponentProps } from './types/ComponentProps'
@@ -57,15 +68,6 @@ export type { WithFoundation } from './types/WithFoundation'
 
 /* Types (Enum) */
 export { default as LayoutHeaderType } from './types/LayoutHeaderType'
-
-/* Hooks */
-export { default as useLayoutDispatch } from './layout/hooks/useLayoutDispatch'
-export { default as useLayoutState } from './layout/hooks/useLayoutState'
-export { default as useSideWidth } from './layout/hooks/useSideWidth'
-export { default as useSidePanelHandler } from './layout/hooks/useSidePanelHandler'
-export { default as useSideViewHandler } from './layout/hooks/useSideViewHandler'
-export { default as useHeader } from './layout/hooks/useHeader'
-export { default as useToast } from './hooks/useToast'
 
 /* Utils */
 export { getRootElement } from './utils/domUtils'
